requirements.txt
altair==4.2.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
astor==0.8.1
attrs==21.4.0
backcall==0.2.0
backports.zoneinfo==0.2.1
base58==2.1.1
bleach==4.1.0
blinker==1.4
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.11
click==8.0.3
colorama==0.4.4
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
entrypoints==0.3
filelock==3.4.2
gitdb==4.0.9
GitPython==3.1.26
huggingface-hub==0.4.0
idna==3.3
importlib-metadata==4.10.1
importlib-resources==5.4.0
ipykernel==6.7.0
ipython==7.31.1
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.18.1
Jinja2==3.0.3
joblib==1.1.0
jsonschema==4.4.0
jupyter-client==7.1.2
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
lxml==4.7.1
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
mistune==0.8.4
nbclient==0.5.10
nbconvert==6.4.1
nbformat==5.1.3
nest-asyncio==1.5.4
notebook==6.4.8
numpy==1.21.5
opencv-python==4.5.5.62
packaging==21.3
pandas==1.3.5
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.0.0
prometheus-client==0.13.1
prompt-toolkit==3.0.26
protobuf==3.19.4
pyarrow==6.0.1
pycparser==2.21
pydeck==0.7.1
Pygments==2.11.2
Pympler==1.0.1
PyMuPDF==1.19.4
pyparsing==3.0.7
pyrsistent==0.18.1
python-dateutil==2.8.2
python-docx==0.8.11
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
pyzmq==22.3.0
regex==2022.1.18
requests==2.27.1
sacremoses==0.0.47
Send2Trash==1.8.0
six==1.16.0
smmap==5.0.0
streamlit==1.5.0
terminado==0.13.1
testpath==0.5.0
tokenizers==0.11.4
toml==0.10.2
toolz==0.11.2
torch==1.10.2+cu113
torchaudio==0.10.2+cu113
torchvision==0.11.3+cu113
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
transformers==4.16.2
typing_extensions==4.0.1
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.8
validators==0.18.2
watchdog==2.1.6
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.2
zipp==3.7.0
PyMuPDF==1.21.0
